add shared bold text function
